Split out router
